move styles into component files
