--- frontend/src/components/Issue/styles.js
@@ -1,48 +1,0 @@
-import {StyleSheet} from 'aphrodite';
-
-const styles = StyleSheet.create({
-  column_wrapper: {
-    padding: "1em 0px",
-    width: "30%",
-    display: "flex",
-    flex: 1,
-    flexDirection: "column",
-    border: "1px solid lightgrey"
-  },
-  issue_style: {
-    cursor: "pointer",
-    textAlign: "left",
-    listStyle: "none",
-    ":hover": {
-      background: "lightgrey"
-    }
-  },
-  select_style: {
-    width: '120px',
-    height: '2em',
-    borderRadius: '3px',
-    overflow: 'hidden',
-    background: 'white'
-  },
-  title: {
-    textAlign: "center"
-  },
-  form: {
-    margin: "10px auto",
-    maxWidth: "600px",
-    display: "grid",
-    alignItems: "baseline",
-    gridTemplateColumns: "100px 1fr",
-    gridGap: "8px"
-  },
-  left_grid: {
-    lineHeight: "1.5em",
-    textAlign: "right",
-    gridColumn: "1 / 2"
-  },
-  right_grid: {
-    gridColumn: "2/3"
-  }
-});
-
-export default styles;
