--- frontend/src/containers/Issue/style.css
@@ -1,30 +1,0 @@
-.wrapper {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  align-items: center;
-}
-.images_wrapper {
-  margin-top: 1em;
-}
-img {
-  padding: 5px;
-}
-form {
-  max-width: 600px;
-  display: grid;
-  grid-template-columns: 100px 1fr;
-  grid-gap: 8px;
-}
-
-label {
-  line-height: 1.5em;
-  text-align: right;
-  grid-column: 1 / 2;
-}
-
-input,
-select,
-button {
-  grid-column: 2 / 3;
-}
